Added RevDate property and column for drawing sheet revision date
Added a new property `RevDate` to the `DrawingSheetModel` class to store the revision date of the drawing sheet. This property is now being set in the `TransmittalViewModel.cs` file with the value of the `SHEET_CURRENT_REVISION_DATE` parameter of the `sheet` object. Additionally, a new column for `RevDate` was added to the `SfDataGrid` control in the `TransmittalView.xaml` file. This column displays the revision date of the drawing sheet, has a header text of "Rev Date", and a width of 100.