Updated UI resources and styles across multiple XAML files
Updated various XAML files with changes including the addition of a new BitmapImage resource in `AppStyles.xaml` for various icons and a new SolidColorBrush resource for WindowForeground. Removed some color resources and added a new style for Window. Commented out the DataContext and applied a new style to the Window from the StaticResource `WindowStyle` in multiple files. Rearranged properties of the Window and set the WindowStyle to `ToolWindow` in `NewRevisionView.xaml` and `RevisionsView.xaml`. Removed FontSize and FontFamily from the Window in `ProgressView.xaml`, `SettingsView.xaml`, and `TransmittalView.xaml`. Set MinWidth and MinHeight in `SettingsView.xaml`. Set WindowStartupLocation to `CenterOwner` and WindowStyle to `SingleBorderWindow` in `StatusView.xaml` and `TransmittalView.xaml`.