Commit: Added Git and System namespaces, updated signing process, and cleaned up unused namespaces
This commit includes several significant changes. The `Nuke.Common.Git` namespace was added to `Build.Installer.cs` and `Build.cs` to provide tools for working with Git repositories. A condition was added to the `Installer` target to only execute if the current Git branch is the main or master branch. The `System` namespace was added to `Build.Sign.cs` to provide fundamental classes and base classes. The signing process was updated to use environment variables for the certificate file and password for better security. Several unused namespaces were removed from `Build.cs`. The `GitRepository` attribute was added to `Build.cs` to inject a `GitRepository` instance into the build. Lastly, the `Solution` attribute in `Build.cs` was updated to generate projects.