Add Master Docs List Report feature
closes #49

Implemented a new feature allowing users to generate and view a Master Documents List Report directly from the UI. This includes the addition of a new method `MasterDocumentsListReport` in the `ArchiveViewModel` class, which interacts with the `Reports` class to prepare and display the report. A corresponding UI button was added to `ArchiveView.xaml`, bound to the new command for report generation. The `Reports` class was extended to include a method `ShowMasterDocumentsListReport`, utilizing a new RDLC report file `MasterDocumentsList.rdlc` as an embedded resource for report layout. Additionally, the RDLC report definition was detailed, including data sources, datasets, layout, and metadata, setting up a comprehensive structure for the report focused on project document summaries.