Add Transmittal_5_Dir and update dependencies
This commit introduces new directories and components for `Transmittal_5_Dir` and `Data_6_Dir`, along with associated resources and files. It includes the addition of new rows for `Data_6_Dir`, `Resources_6_Dir`, and `Transmittal_5_Dir`, as well as new files like `TemplateDatabase.tdb_6` and `TransmittalParameters.txt_6`. Additionally, several outdated components related to libraries such as `CommunityToolkit.Mvvm`, `Dapper`, and `Humanizer` have been removed, indicating a cleanup and update of dependencies to support new features in the application.